Structure change for components
